[LI-HOTFIX] Disable the logging of UpdateMetadata live brokers in the LiCombinedControlRequest (#215)
TICKET = N/A
LI_DESCRIPTION =
Initially, when developing the LiCombinedControlRequest feature, we turned on logging
of UpdateMetadata live brokers in the LiCombinedControlRequest to ensure the correctness of the
feature. Now that the feature has proven to be stable, it's better to turn off the logging
to make the log files smaller.
EXIT_CRITERIA = This change has the same exit criteria as the LiCombinedControlRequest feature

* Reword the comment