[LI-TIERING] Remove acquiring write lock when closing the Remote Log Manager
Summary:
RLM needs a write lock to close itself. But, the read lock is held by other threads which are under the infinite loop:
  # Read lock held by ConsumerManager won't be returned when it's in `waitTillConsumptionCatchesUp` method.
  # Read lock held by ProducerManager won't be released when it's in `publishMessage` method.

Delivery timeout is set to 2 minutes by default. So, once the timeout exceeds the read lock held by the producer manager gets released.

Test Plan:
Prior to this patch, the TransactionsTestWithTieredStore test takes 20+ minutes to run, now it takes 8 minutes. The normal variant takes 5+ minutes to run the test.

Out of 19 tests, 4 of them each wait for around 35 seconds to get the log segments uploaded to remote storage and removed from the local-log storage physically . This constraint is due to the initial task delay to run the log cleaner thread being set to 30 seconds. (Compared to normal one, extra time required is 35 x 4 = 140 seconds).

The wait timeout is increased to 40 seconds for reliability.

Reviewers: satishd, prateeka, kjayaraman

Reviewed By: satishd

JIRA Issues: DKAFC-1588

Differential Revision: https://code.uberinternal.com/D6928497