[LI-HOTFIX] Add fetchState information in NoOffsetForPartitionException (#285)
TICKET = LIKAFKA-40747
LI_DESCRIPTION =
A potential reason for NoOffsetForPartitionException is the transition of FetchState for partitions to any state other than INITIALIZING (e.g. offset_reset). However, there are not enough logs on the client side to see if that was the case. Add logs in the exception message to track the actual fetch state of partitions when a NoOffsetForPartitionException is thrown.

(cherry picked from commit aaf8c8d76f2a5d3eced83e8bf6265d7243d6278b)