[LI-HOTFIX] Merge controller requests 2/4: adding utilities to merge controller requests (#140)
TICKET = N/A
LI_DESCRIPTION =
This is the 2nd PR to implement the feature of control request merging laid out in the design doc
https://docs.google.com/document/d/1PZAOwpw09tVDeuSP0OBVhB6dbgz7C7dQU6bRkePQ8qg/edit

Specifically, this PR includes
1. adding the ControllerRequestMerger for merging controller requests
2. adding methods in existing classes so that their fields can be accessed.
This includes AbstractControlRequest, LeaderAndIsrRequest, UpdateMetadataRequest, StopReplicaRequest
3. adding a utility class LiCombinedControlTransformer to convert the data format between
the LiCombinedControl request/response and the existing controller request types

EXIT_CRITERIA = It's unclear whether this feature should be put in upstream apache kafka yet. If we decide to put this change upstream, the exit criteria will be when it gets merged upstream and pulled in a part of a release.